requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
click==8.3.0
ecdsa==0.19.1
fastapi==0.117.1
graphql-core==3.2.6
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
lia-web==0.2.3
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.9
pydantic_core==2.33.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
strawberry-graphql==0.282.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
